Refactoring to prepare for examples of nesting
* Removing chattery logs
* Pull useful methods into a spec support module
