+++ lib/samvera/nesting_indexer/version.rb
@@ -1,5 +1,5 @@
 module Samvera
   module NestingIndexer
-    VERSION = "0.7.0".freeze
+    VERSION = "0.8.0".freeze
   end
 end
